requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bbox==0.8.3
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
cloudpickle==1.2.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.10
dask==2.3.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
gradescope-utils==0.4.0
graphviz==0.10.1
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
ipdb==0.13.4
ipykernel==5.1.2
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
loguru==0.3.2
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
multipledispatch==0.6.0
nbconvert==5.6.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.3
nose==1.3.7
notebook==5.7.4
numpy==1.16.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
opencv-contrib-python==4.1.2.30
opencv-python==3.4.5.20
padme==1.1.1
pandocfilters==1.4.2
parso==0.5.1
pbr==5.4.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.1.0
plainbox==0.25
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.0.0
ptyprocess==0.6.0
pyasn1==0.4.2
pycairo==1.16.2
pycocotools==2.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyquaternion==0.9.5
pyRFC3339==1.0
pyrr==0.10.0
pyrsistent==0.15.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.0
python-debian==0.1.32
python-pcl==0.3.0a1
pytz==2018.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
qtconsole==4.5.5
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.2
scipy==1.3.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
ssh-import-id==5.7
stevedore==1.30.1
subprocess32==3.5.4
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
torchfile==0.1.0
tornado==6.0.3
traitlets==4.3.2
transforms3d==0.3.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
websocket-client==0.56.0
widgetsnbextension==3.5.1
xkit==0.0.0
XlsxWriter==0.9.6
zope.interface==4.3.2
